feat: Make add() safe
As per discussion in #255, the add() function can be made safe with no
other modifications to the API. This allows the `polling` crate to be
used entirely safely again.

The main changes to the public API are add() and add_waitable() being
made safe. This is a non-breaking change, as it is allowed to move a
function from unsafe to safe without a breaking change.

I don't like `AsRawSource` still being in the public API, but that
can't really be helped at this point. If we do another breaking change
in the future this can be fixed.

Signed-off-by: John Nunley <dev@notgull.net>